Major update to conseno_enhanced conversions_variable.tpl
Now, the consent field is not only tied to cookie but you can add any GTM variable into this field that has either true or false (boolean or string) values. This means that you don’t even need to generate a 1st party cookie anymore (if you don’t want to) but you can simply take in-line consent decision value from the dataLayer variable as well. But, if the in-line consent decision value is not pushed to the dataLayer along with an email after the form submit, you can still generate a 1st party cookie and add a 1st party cookie variable into this “Select the in-line consent variable” field. Also, in some cases, the in-line consent value might not be true/false by default (it may be, for example, accept/decline) meaning that you may need adjust it with the Custom JavaScript variable before you enter it to the  “Select the in-line consent variable” field. I think, now this template technically makes much more sense and is more flexible. Now, it also won’t require cookie reading permissions anymore, etc.